@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.8.1
+  sdk: ^3.9.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -49,10 +49,11 @@ dependencies:
   zego_uikit_signaling_plugin: ^2.8.18
   zego_uikit: ^2.28.28
   iconsax: ^0.0.8
-  record: ^5.0.0
+  record: ^6.1.0
   path_provider: ^2.1.3
   bugsnag_flutter: ^4.2.0
   just_audio: ^0.10.4
+# pubspec.yaml ga qo‘shing (vaqtinchalik)
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -76,8 +77,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
+  assets:
+     - assets/images/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
