@@ -1,3 +1,4 @@
+// custom_audio_bubble.dart
 import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:just_audio/just_audio.dart';
@@ -8,6 +9,7 @@ class CustomAudioBubble extends StatefulWidget {
   final Color bubbleColor;
   final Color textColor;
   final Color accentColor;
+  final String uniqueId; // Yangi: har bir bubble uchun unikal ID
 
   const CustomAudioBubble({
     required this.audioUrl,
@@ -15,6 +17,7 @@ class CustomAudioBubble extends StatefulWidget {
     required this.bubbleColor,
     required this.textColor,
     required this.accentColor,
+    required this.uniqueId,
     Key? key,
   }) : super(key: key);
 
@@ -29,6 +32,8 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
   late final List<double> _waveHeights;
 
   bool _isPlaying = false;
+  bool _isLoading = true;
+  bool _hasError = false;
   Duration _duration = Duration.zero;
   Duration _position = Duration.zero;
   double _speed = 1.0;
@@ -42,22 +47,37 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
       duration: const Duration(milliseconds: 700),
     );
     _waveHeights = List.generate(16, (i) => 4 + Random().nextDouble() * 12);
-    _initPlayer();
+    _initPlayer(); // Faqat bir marta
   }
 
   Future<void> _initPlayer() async {
+    if (!mounted) return;
+
+    setState(() {
+      _isLoading = true;
+      _hasError = false;
+    });
+
     try {
       await _player.setUrl(widget.audioUrl);
+      _duration = _player.duration ?? Duration.zero;
 
+      // Duration stream
       _player.durationStream.listen((dur) {
-        if (dur != null) setState(() => _duration = dur);
+        if (!mounted || dur == null) return;
+        setState(() => _duration = dur);
       });
 
+      // Position stream
       _player.positionStream.listen((pos) {
+        if (!mounted) return;
         setState(() => _position = pos);
       });
 
+      // Player state stream
       _player.playerStateStream.listen((state) {
+        if (!mounted) return;
+
         final playing = state.playing;
         final completed = state.processingState == ProcessingState.completed;
 
@@ -69,27 +89,49 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
           _player.pause();
           _player.seek(Duration.zero);
           _waveController.stop();
+          setState(() => _position = Duration.zero);
         } else if (playing) {
           _waveController.repeat(reverse: true);
         } else {
           _waveController.stop();
         }
       });
+
+      setState(() => _isLoading = false);
     } catch (e) {
       debugPrint('Audio load error: $e');
+      if (mounted) {
+        setState(() {
+          _isLoading = false;
+          _hasError = true;
+        });
+      }
+    }
+  }
+
+  @override
+  void didUpdateWidget(covariant CustomAudioBubble oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    // Faqat URL o'zgarsa qayta yuklash
+    if (oldWidget.audioUrl != widget.audioUrl) {
+      _player.stop();
+      _initPlayer();
     }
   }
 
   @override
   void dispose() {
+    _player.stop();
+    _waveController.stop();
     _waveController.dispose();
     _player.dispose();
     super.dispose();
   }
 
-  String _formatMinutes(Duration d) {
-    final m = d.inMinutes.remainder(60).toString().padLeft(2, '0');
-    return '$m:${d.inSeconds.remainder(60).toString().padLeft(2, '0')}';
+  String _formatDuration(Duration d) {
+    final minutes = d.inMinutes.remainder(60).toString().padLeft(2, '0');
+    final seconds = d.inSeconds.remainder(60).toString().padLeft(2, '0');
+    return '$minutes:$seconds';
   }
 
   @override
@@ -123,9 +165,11 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
         child: Row(
           mainAxisSize: MainAxisSize.min,
           children: [
-            // ▶️ Play / Pause
+            // Play / Pause
             InkWell(
-              onTap: () async {
+              onTap: _isLoading || _hasError
+                  ? null
+                  : () async {
                 if (_isPlaying) {
                   await _player.pause();
                 } else {
@@ -136,7 +180,18 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
               child: CircleAvatar(
                 radius: 22,
                 backgroundColor: widget.accentColor.withOpacity(0.1),
-                child: Icon(
+                child: _isLoading
+                    ? const SizedBox(
+                  width: 18,
+                  height: 18,
+                  child: CircularProgressIndicator(
+                    strokeWidth: 2,
+                    color: Colors.white,
+                  ),
+                )
+                    : _hasError
+                    ? const Icon(Icons.error, color: Colors.red, size: 20)
+                    : Icon(
                   _isPlaying ? Icons.pause : Icons.play_arrow,
                   color: widget.accentColor,
                   size: 28,
@@ -146,15 +201,14 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
 
             const SizedBox(width: 10),
 
-            // 🔊 Animated Waveform
+            // Waveform
             Expanded(
               child: AnimatedBuilder(
                 animation: _waveController,
                 builder: (context, _) {
-                  final value = _waveController.value;
                   return CustomPaint(
                     painter: _WaveformPainter(
-                      value: value,
+                      value: _waveController.value,
                       progress: progress,
                       accentColor: widget.accentColor,
                       baseHeights: _waveHeights,
@@ -167,9 +221,9 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
 
             const SizedBox(width: 12),
 
-            // ⏱ Duration
+            // Duration
             Text(
-              _formatMinutes(_duration),
+              _formatDuration(_duration),
               style: TextStyle(
                 fontSize: 12,
                 color: widget.textColor.withOpacity(0.8),
@@ -179,20 +233,16 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
 
             const SizedBox(width: 8),
 
-            // 🔁 Speed toggle
+            // Speed
             InkWell(
-              onTap: () {
+              onTap: _isLoading || _hasError
+                  ? null
+                  : () {
                 setState(() {
-                  if (_speed == 1.0)
-                    _speed = 1.5;
-                  else if (_speed == 1.5)
-                    _speed = 2.0;
-                  else
-                    _speed = 1.0;
+                  _speed = _speed == 1.0 ? 1.5 : _speed == 1.5 ? 2.0 : 1.0;
                   _player.setSpeed(_speed);
                 });
               },
-              borderRadius: BorderRadius.circular(6),
               child: Container(
                 padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 3),
                 decoration: BoxDecoration(
@@ -201,11 +251,7 @@ class _CustomAudioBubbleState extends State<CustomAudioBubble>
                 ),
                 child: Text(
                   '${_speed.toStringAsFixed(1)}x',
-                  style: const TextStyle(
-                    fontSize: 11,
-                    fontWeight: FontWeight.w600,
-                    color: Colors.black87,
-                  ),
+                  style: const TextStyle(fontSize: 11, fontWeight: FontWeight.w600),
                 ),
               ),
             ),
@@ -243,10 +289,9 @@ class _WaveformPainter extends CustomPainter {
       final x = i * (barWidth + gap);
       final baseHeight = baseHeights[i];
       final animHeight = baseHeight + sin((value * pi * 2) + i) * 4;
-      final barHeight = animHeight.clamp(4, 28);
+      final barHeight = animHeight.clamp(4.0, 28.0);
 
-      final progressRatio = (i / totalBars);
-      final isPlayed = progressRatio < progress;
+      final isPlayed = (i / totalBars) < progress;
 
       paint.color = isPlayed
           ? accentColor.withOpacity(0.95)
@@ -255,7 +300,7 @@ class _WaveformPainter extends CustomPainter {
       final y = (size.height - barHeight) / 2;
       canvas.drawRRect(
         RRect.fromRectAndRadius(
-          Rect.fromLTWH(x, y, barWidth, barHeight.toDouble()),
+          Rect.fromLTWH(x, y, barWidth, barHeight),
           const Radius.circular(3),
         ),
         paint,
@@ -264,7 +309,5 @@ class _WaveformPainter extends CustomPainter {
   }
 
   @override
-  bool shouldRepaint(_WaveformPainter oldDelegate) {
-    return oldDelegate.value != value || oldDelegate.progress != progress;
-  }
+  bool shouldRepaint(covariant CustomPainter oldDelegate) => true;
 }
